configuration update

+++ confs/create_mysql_db.sql
@@ -1,12 +1,26 @@
 DROP SCHEMA IF EXISTS `servermail`;
 CREATE SCHEMA `servermail`;
-GRANT USAGE ON *.* TO 'usermail'@'%';
-DROP USER 'usermail'@'%';
+USE 'servermail';
+DROP PROCEDURE IF EXISTS servermail.drop_user_if_exists ;
+DELIMITER $$
+CREATE PROCEDURE servermail.drop_user_if_exists()
+BEGIN
+  DECLARE foo BIGINT DEFAULT 0 ;
+  SELECT COUNT(*)
+  INTO foo
+    FROM mysql.user
+      WHERE User = 'usermail' and  Host = '%';
+    IF foo > 0 THEN
+      DROP USER 'usermail'@'%' ;
+  END IF;
+END ;$$
+DELIMITER ;
+CALL servermail.drop_user_if_exists() ;
+DROP PROCEDURE IF EXISTS servermail.drop_users_if_exists ;
 CREATE USER 'usermail'@'%';
 SET PASSWORD FOR 'usermail'@'%' = PASSWORD('%MAILUSER_PSW%');
 GRANT SELECT ON `servermail`.* TO 'usermail'@'%';
 FLUSH PRIVILEGES;
-USE 'servermail';
 CREATE TABLE `virtual_domains` (
 	`id` int(11) NOT NULL auto_increment,
 	`name` varchar(50) NOT NULL, PRIMARY KEY (`id`)) ENGINE=InnoDB DEFAULT CHARSET=utf8;
